+++ rqueue-core/src/main/java/com/github/sonus21/rqueue/models/request/DataTypeRequest.java
@@ -14,18 +14,21 @@
  *
  */
 
-package com.github.sonus21.task.executor;
+package com.github.sonus21.rqueue.models.request;
 
-import lombok.AllArgsConstructor;
-import lombok.Data;
+import com.github.sonus21.rqueue.models.SerializableBase;
+import javax.validation.constraints.NotEmpty;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.Setter;
+import lombok.ToString;
 
-@Data
-@AllArgsConstructor
+@Getter
+@Setter
 @NoArgsConstructor
-public class Email {
-
-  private String email;
-  private String subject;
-  private String content;
+@ToString
+@EqualsAndHashCode(callSuper = true)
+public class DataTypeRequest extends SerializableBase {
+  @NotEmpty private String name;
 }