+++ rqueue-core/src/main/java/com/github/sonus21/rqueue/models/request/DateViewRequest.java
@@ -1,0 +1,45 @@
+/*
+ *  Copyright 2021 Sonu Kumar
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *         https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and limitations under the License.
+ *
+ */
+
+package com.github.sonus21.rqueue.models.request;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.github.sonus21.rqueue.models.SerializableBase;
+import com.github.sonus21.rqueue.models.enums.DataType;
+import javax.validation.constraints.NotEmpty;
+import javax.validation.constraints.NotNull;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import lombok.ToString;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@ToString
+@EqualsAndHashCode(callSuper = true)
+public class DateViewRequest extends SerializableBase {
+  private @NotNull DataType type;
+  private @NotEmpty String name;
+  private  String key;
+
+  @JsonProperty("page")
+  private int pageNumber = 0;
+
+  @JsonProperty("count")
+  private int itemPerPage = 20;
+}