+++ rqueue-spring-boot-reactive-example/src/main/java/com/github/sonus21/task/executor/Controller.java
@@ -17,49 +17,57 @@
 package com.github.sonus21.task.executor;
 
 import com.github.sonus21.rqueue.core.ReactiveRqueueMessageEnqueuer;
-import lombok.NonNull;
-import lombok.RequiredArgsConstructor;
+import java.util.UUID;
+import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import reactor.core.publisher.Mono;
 
 @RestController
-@RequiredArgsConstructor(onConstructor = @__(@Autowired))
+@AllArgsConstructor(onConstructor = @__(@Autowired))
 @Slf4j
 public class Controller {
 
-  private @NonNull
-  final ReactiveRqueueMessageEnqueuer reactiveRqueueMessageEnqueuer;
+  private final ReactiveRqueueMessageEnqueuer reactiveRqueueMessageEnqueuer;
 
-  @Value("${email.queue.name}")
-  private String emailQueueName;
-
-  @Value("${invoice.queue.name}")
-  private String invoiceQueueName;
-
-  @Value("${invoice.queue.delay}")
-  private Long invoiceDelay;
-
-  @GetMapping("email")
-  public Mono<String> sendEmail(
-      @RequestParam String email, @RequestParam String subject, @RequestParam String content) {
-    log.info("Sending email");
-    Mono<String> mono = reactiveRqueueMessageEnqueuer
-        .enqueue(emailQueueName, new Email(email, subject, content));
-    return mono.zipWith(Mono.just("Please check your inbox!"), (a, b) -> b);
+  @GetMapping(value = "/push")
+  public Mono<String> push(
+      String q,
+      String msg,
+      @RequestParam(required = false) Integer numRetries,
+      @RequestParam(required = false) Long delay) {
+    Mono<String> data;
+    if (numRetries == null && delay == null) {
+      data = reactiveRqueueMessageEnqueuer.enqueue(q, msg);
+    } else if (numRetries == null) {
+      data = reactiveRqueueMessageEnqueuer.enqueueIn(q, msg, delay);
+    } else {
+      data = reactiveRqueueMessageEnqueuer.enqueueInWithRetry(q, msg, numRetries, delay);
+    }
+    log.info("Message {}", msg);
+    return data;
   }
 
-  @GetMapping("invoice")
-  public Mono<String> generateInvoice(@RequestParam String id, @RequestParam String type) {
-    log.info("Generate invoice");
-    Mono<String> mono = reactiveRqueueMessageEnqueuer
-        .enqueueIn(invoiceQueueName, new Invoice(id, type), invoiceDelay);
-    return mono
-        .zipWith(Mono.just("Invoice would be generated in " + invoiceDelay + " milliseconds"),
-            (a, b) -> b);
+  @GetMapping("job")
+  public Mono<String> sendJobNotification() {
+    Job job = new Job();
+    job.setId(UUID.randomUUID().toString());
+    job.setMessage("Hi this is " + job.getId());
+    Mono<String> mono = reactiveRqueueMessageEnqueuer.enqueue("job-queue", job);
+    log.info("{}", job);
+    return mono;
+  }
+
+  @GetMapping("job-delay")
+  public Mono<String> sendJobNotificationWithDelay() {
+    Job job = new Job();
+    job.setId(UUID.randomUUID().toString());
+    job.setMessage("Hi this is " + job.getId());
+    Mono<String> mono = reactiveRqueueMessageEnqueuer.enqueueIn("job-queue", job, 2000L);
+    log.info("{}", job);
+    return mono;
   }
 }